Re-organize projects data
